<feat>: Set up database and finish implement Dao, Model
Check the API of Dao interface and see the sample usage in the DaoTest
package.
For database, the movies in the movies.txt must be distinct. But One movie
can have multiple bookings in the bookings.txt(expired bookings will be
deleted automatically). The price of movie is depends on the screen and
it is store as "Screen Size#Price" in the movie.txt. The status 0 in the
gift_cards.txt means the card has expired.